requirements.txt
python-telegram-bot==22.6
google-generativeai==0.8.6
